back-end-101/task-2/pyproject.toml
[tool.poetry]
name = "your_project_name"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"  # или ваша версия Python
requests = "^2.26"  # версия requests
numpy = "^1.21"  # версия numpy
python-dotenv = "^0.19"  # версия python-dotenv

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"